requirements.txt
# Zero-Loss Circuit Breaker
# Autonomous Payment Dispute Resolution System

langgraph>=0.0.40
langchain-openai>=0.1.0
pydantic>=2.0
python-dotenv>=1.0.0
rich>=13.0.0
openai>=1.0.0

